requirements.txt
clickhouse-connect>=0.6.0
python-dotenv>=1.0.0
schedule>=1.2.0
psutil>=5.9.0